Time: 42 ms (83.87%), Space: 21.8 MB (61.97%) - LeetHub

+++ 2101-detonate-the-maximum-bombs/2101-detonate-the-maximum-bombs.cpp
@@ -15,7 +15,7 @@ public:
         vector<int> graph[n];
 
         for(int i=0; i<n; i++){
-            int x = bombs[i][0]; int y = bombs[i][1]; int radius = bombs[i][2];
+            long long x = bombs[i][0]; long long y = bombs[i][1]; long long radius = bombs[i][2];
             for(int j=0; j<n; j++){
                 if(i == j) continue;
                 else{
